refactor: move dense union into vortex-spatial
