--- encodings/dense-union/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "vortex-dense-union"
-authors = { workspace = true }
-categories = { workspace = true }
-description = "Dense union encoding for Vortex arrays"
-edition = { workspace = true }
-homepage = { workspace = true }
-include = { workspace = true }
-keywords = { workspace = true }
-license = { workspace = true }
-readme = "README.md"
-repository = { workspace = true }
-rust-version = { workspace = true }
-version = { workspace = true }
-
-[dependencies]
-prost = { workspace = true }
-vortex-array = { workspace = true }
-vortex-buffer = { workspace = true }
-vortex-error = { workspace = true }
-vortex-mask = { workspace = true }
-vortex-session = { workspace = true }
-
-[dev-dependencies]
-divan = { workspace = true }
-rstest = { workspace = true }
-vortex-array = { workspace = true, features = ["_test-harness"] }
-
-[lints]
-workspace = true
-
-[[bench]]
-name = "take"
-harness = false
