--- encodings/dense-union/README.md
@@ -1,3 +1,0 @@
-# Vortex Dense Union
-
-An external dense physical encoding for Vortex's logical `DType::Union`.
