+++ vortex-spatial/src/dense_union/compute/mod.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+//! Compute kernels for dense unions.
+
 mod filter;
 mod mask;
 mod slice;
